[refactor/fix] 过滤器重构(待完善), 修复各种错误

+++ numeron-annotation/src/main/java/com/erzbir/numeron/annotation/Message.java
@@ -34,5 +34,5 @@ public @interface Message {
 
     ConcurrencyKind concurrency() default ConcurrencyKind.CONCURRENT;
 
-    int id() default 0;
+    long id() default 0;
 }
